+++ src/app/api/mycoupon/route.ts
@@ -8,21 +8,21 @@ export async function GET(req: NextRequest) {
   const hasCookies = cookies().has('AUTH_TOKEN')
   try {
     const requestHeaders = new Headers(req.headers)
-    // console.log('cookie', cookies().getAll())
+
     if (hasCookies) {
       requestHeaders.set('Authorization', `Bearer ${authToken}`)
     }
     console.log('쿠폰 목록 !! Authorization', requestHeaders)
 
-    const res = await fetch(`${baseURL}/members/mypage/coupons`, {
+    const res = await fetch(`${baseURL}/items/getcouponList`, {
       method: 'GET',
       headers: requestHeaders,
-      // headers: commonHeader,
+     
     })
     const response = await res.json()
     return NextResponse.json(response)
 
-    // return res
+  
   } catch (error) {
     console.log('couponList 에러', error)
     return NextResponse.json({ msg: 'error' })
