+++ src/components/feature/order/orderForm.tsx
@@ -12,7 +12,7 @@ import CheckModal from '@/components/common/modal/checkModal'
 import OrderAddress from '@/components/feature/order/orderAddress'
 import OrderBill from '@/components/feature/order/orderBill'
 import OrderContainer from '@/components/feature/order/orderContainer'
-import OrderCoupon from '@/components/feature/ordercoupon/orderCoupon'
+import OrderCoupon from '@/components/feature/order/orderCoupon'
 import OrderPaymentButton from '@/components/feature/order/orderPaymentButton'
 import OrderPaymentMethod from '@/components/feature/order/OrderPaymentMethod'
 import OrderReserves from '@/components/feature/order/orderReserves'
